Update SignUpViewModelTest - added more cases
Through unit tests, the validation of fields (name, email, password and password repetition) is evaluated in the Sign Up View Model.
The cases evaluated are:
complete fields, incomplete, cases of errors in the mail, weak passwords that do not meet the conditions, password and its repetition do not match, etc.